refactor: using livedata to track locale
previously, i set changelanguageviewmodel's variable inside fragment, which is not a good practice. so, i change the approach with saving locale value and wrap it with livedata so i can observe it in the fragment. i also change how the observer will be called, so that the fragment doesn't create multiple instances of viewmodel.